Add git hash image to About box, available by tap on app name.
Switch from using a .gitignored file containing revision information to a script that adds that info to the Info.plist of the build product.

Co-opt MITBuildInfo file name for a helper class to access new build information and to generate the hash image.

Update how we name the main Core Data sqlite file to use git, as that previously depended on SVN revision numbers.