Stop spamming requests for the rooms for a building.
CoreData's managedObjectContexts are not meant to be shared across threads, so CoreDataManager automatically uses an appropriate moc for the current thread. Blocks on GCD queues are not guaranteed to execute on the same thread, so things like saving asynchronously and then synchronously checking for a value won't work, even on the same queue.

This code used to keep returning stale data for location.roomsUpdated until the stalenessInterval of the moc had been reached. So even though rooms might have just been updated and saved to the persistentstore, that data is not reflected in an object which was already in the current thread's moc's cache. By adding a call to -refreshObject:mergeChanges:, the location is turned into a fault and the moc then reloads that object from disk when location.roomsUpdated is called.