Correct sizing for the My Account table cells
Use hard coded values for the width when calculating the desired height of the cells. While this may (and probably) will break in the future, it should be much easier to fix than figuring out exactly how the sizing calculations should be changed.

To Do: Look into using a generic UITableViewCell (setup in a similar manner), layout its subviews and pull the frame of the contentView