IPHONEAPP-596. The academic year was being calculated incorrectly in the Holidays calendar.
From the documentation on NSDateComponents: "When a new instance of NSDateComponents is created the date components are set to NSUndefinedDateComponent."
NSUndefinedDateComponent = 0x7fffffff, so the date was always considered to be later than the 6th month of the year. As it turns out this file on GitHub is the 2nd Google result for "nsdate fiscal year".