IPHONEAPP-600
Add the ability to request arbitrary data using the MobileRequestOperation. The requested data will be returned along with the content type reported by the server. At of this commit, the server does not return proper Content-Type information for JSON data so we have to attempt to 'guess' if what we have is JSON or not. Do not remove the warning directive in MobileRequestOperation until this bug has been fixed.