Take advantage of the fact we're using blocks and collapse the MobileRequestOperation response handling chain in the Links module.
Get rid of -reloadTableView. It's a one line replacement for a one line call.

Use GCD instead of NSTimer to show the loading view, reducing the number of methods needed and making the code more readable.

Get rid of the unused requestWasDispatched property.

Add a linkCachePath helper method to make sure we never typo that path creation code.