Build profile cleanup
Revamp the build profiles. There are now 3 profiles (Debug,Release,TestFlight) and 2 schemes. The Release and TestFlight profiles are identical except that TestFlight will use the iOS Developer cert to sign the build and Release requires an iOS distribution cert. The 'TestFlight' profile is only used when executing the 'Archive' action in the 'MIT Mobile (Development)' scheme

Signed-off-by: Blake Skinner <bskinner@mit.edu>