Adding autoload of more stories to news categories.  #571
Added storyRefreshInProgress to make sure we don't signal a get more stories while we are refreshing.

We do not want to refresh loading cell if the cell is not visible or if the tableview is moving or there is no error.
If the tableview is moving we have already updated the status of the cell because we called this ourselves when we came across the cell.

If they are clicking the cell themselves the table is not moving and the cell is shown.

But if there is an error we want to change the cell.

References #571